// src/shared/interfaces/spreadsheet-structure.interface.ts
import {SpreadsheetInfo} from './spreadsheet-info.interface';

export interface SpreadsheetStructure {
  IRM_BRA: SpreadsheetInfo;
  IRM_SSC: SpreadsheetInfo;
  SOPORTE_RECOVERY: SpreadsheetInfo;
  FUP_T90: SpreadsheetInfo;
  OPEN_PURCHASES: SpreadsheetInfo;
  OPEN_REPAIRS: SpreadsheetInfo;
  REPAIR_STATUSES: SpreadsheetInfo;
  PURCHASE_STATUSES: SpreadsheetInfo;
}